refactor: fetchDailyRecords 함수에서 useQuery 훅을 사용한 useGetDailyRecords로 변경
